Refactoring (#1):     - Logging -> Log
