+++ andromeda/BUILD.bazel
@@ -11,6 +11,6 @@ cc_library(
     deps = [
         "@andromeda//andromeda/core",
         "@andromeda//andromeda/core:instance",
-        "@andromeda//andromeda/core/system/logging:log",
+        "@andromeda//andromeda/core/system/log",
     ],
 )
