+++ andromeda/core/link.hpp
@@ -4,7 +4,7 @@
 
 #include "andromeda/core/core.hpp"
 #include "andromeda/core/instance.hpp"
-#include "andromeda/core/system/logging/log.hpp"
+#include "andromeda/core/system/log/log.hpp"
 
 extern std::unique_ptr<Andromeda::Instance> Andromeda::instantiate();
