+++ andromeda/core/system/log/log.cpp
